-- 5-5. 답변 추가 및 목록 기능 구현 - 답변를 하기 위해 답변 테이블 Answer.jave 클래스 생성 -- src/main/java/net/dndlti/domain/Answer.java //데이터베이스와 연동하는 클래스라는 것을 표시하기 위해 //@Entity어노테이션 사용 //클래스 이름으로 H2 데이터베이스의 Answer 테이블 생성됨 //- jpa 라이브러리 사용으로 가능함 //답변(하기) 테이블 생성 @Entity public class Answer {
  //id 컬럼을 primary key 로 사용하고 숫자를 자동증가하는 방식으로 타입 설정
  //@GeneratedValue - 자동으로 1씩 증가
  //데이터형 Long 을 사용해야 함 - long 타입과 동일하지 않음
  @Id
  @GeneratedValue
  private Long id;

  //User 객체와 관계를 맺기 위해 필드 생성
  //answer 와 user 객체는 ManyToOne 관계이다.
  //User 객체에 대한 외래키를 지정하기 위해 @JoinColumn 어노테이션 사용하여
  //foreign 속성을 사용할 수 있다.
  //User 테이블의 기본키인 id 컬럼을 참조하는
  //User 객체의 인스턴스 객체 writer 생성
  @ManyToOne
  @JoinColumn(
  foreignKey = @ForeignKey(name="fk_answer_writer"))
  private User writer;

  //하나의 질문(question) 에는 여러 개의 답변(answer)이 달리기 때문에
  //answer 와 question 객체는 ManyToOne 관계이다.
  //question 객체에 대한 외래키를 지정하기 위해 @JoinColumn 어노테이션 사용하여
  //foreignKey 속성을 사용할 수 있다.
  //Question 테이블의 기본키인 id 컬럼을 참조하는
  //Question 객체의 인스턴스 객체 question 생성
  @ManyToOne
  @JoinColumn(
  foreignKey=@ForeignKey(name="fk_answer_to_question"))
  private Question question;

 private String contents;

  /*5-1 질문하기 입력 시간을 표시하는 컬럼 추가*/
  //LocalDateTime - Java 1.8 부터 지원하는
  //다수의 스레드들로부터 안전하게 접근할 수 있고 변경이 불가능한 final class
  private LocalDateTime createDate;

  public Answer() {
  }

  .........................................

+++ src/main/java/net/dndlti/web/QuestionController.java
@@ -21,7 +21,7 @@ import net.dndlti.domain.User;
 @RequestMapping("/questions")
 public class QuestionController {
   
-  //데이터베이스 테이블과의 연동
+  //QuestionRepository 인터페이스는 데이터베이스 테이블 Question 과 연동한다.
   //@Autowired 어노테이션은 
   //스프링에서 자동으로 의존관계를 타입에 맞춰서 연결하도록 시킴
   //Marks a constructor, field, setter method or 
@@ -81,19 +81,52 @@ public class QuestionController {
   }
   
   //질문 수정 폼으로 이동
+  //-- 세션 로그인한 질문 작성자만이 질문 수정하도록 구현
   @GetMapping("/{id}/form")
   public String updateForm(
-  @PathVariable Long id, Model model) {
-    model.addAttribute(
-    "question", questionRepository.findOne(id));
+  @PathVariable Long id, Model model, HttpSession session) {
+    //세션로그인 여부를 확인할 수 있는 메소드
+    System.out.println(
+    "나의 ~ !HttpSessionUtils.isLoginUser(session): " 
+    + !HttpSessionUtils.isLoginUser(session));
+    if (!HttpSessionUtils.isLoginUser(session)) {
+      /*css,js,bootstrap,jquery 등의 자원을 제대로 읽어오기 위해 수정*/
+      return "redirect:/users/loginForm";
+    }
+    //세션로그인을 한 User 객체인지를 확인
+    User loginUser = 
+    HttpSessionUtils.getUserFromSession(session);
+    
+    Question question = questionRepository.findOne(id);
+    System.out.println(
+    "나의 ~ !question.isSameWriter(loginUser): " 
+    + !question.isSameWriter(loginUser));
+    if (!question.isSameWriter(loginUser)) {
+      /*css,js,bootstrap,jquery 등의 자원을 제대로 읽어오기 위해 수정*/
+      return "redirect:/users/loginForm";
+    }
+    model.addAttribute("question", question);
     return "/qna/updateForm";
   }
   
-  //질문 수정을 한 후 해당 질문이 표시되도록 페이지 이동
+  //질문 수정을 한 후 해당 질문이 표시되도록 페이지 이동 
+  //로그인 사용자중에 질문 작성자만이 질문 수정할 수 있도록 구현
   @PutMapping("/{id}")
   public String update(
-  @PathVariable Long id, String title, String contents) {
+  @PathVariable Long id, String title, String contents,
+  HttpSession session) {
+    if (!HttpSessionUtils.isLoginUser(session)) {
+      /*css,js,bootstrap,jquery 등의 자원을 제대로 읽어오기 위해 수정*/
+      return "redirect:/users/loginForm";
+    }
+    User loginUser = 
+    HttpSessionUtils.getUserFromSession(session);
     Question question = questionRepository.findOne(id);
+    if (!question.isSameWriter(loginUser)) {
+      /*css,js,bootstrap,jquery 등의 자원을 제대로 읽어오기 위해 수정*/
+      return "redirect:/users/loginForm";
+    }
+    
     question.update(title, contents);
     questionRepository.save(question);
     return String.format("redirect:/questions/%d", id);
@@ -101,7 +134,19 @@ public class QuestionController {
   
   //질문 작성자가 질문 삭제 
   @DeleteMapping("/{id}")
-  public String delete(@PathVariable Long id) {
+  public String delete(@PathVariable Long id,
+  HttpSession session) {
+    if (!HttpSessionUtils.isLoginUser(session)) {
+      /*css,js,bootstrap,jquery 등의 자원을 제대로 읽어오기 위해 수정*/
+      return "redirect:/users/loginForm";
+    }
+    User loginUser = HttpSessionUtils.getUserFromSession(session);
+    Question question = questionRepository.findOne(id);
+    if (!question.isSameWriter(loginUser)) {
+      /*css,js,bootstrap,jquery 등의 자원을 제대로 읽어오기 위해 수정*/
+      return "redirect:/users/loginForm";
+    }
+    
     questionRepository.delete(id);
     return "redirect:/";
   }
